Move interaction tracking in own directory
