+++ packages/core/src/rum/instrumentation/interactionTracking/DdRumUserInteractionTracking.tsx
@@ -6,10 +6,10 @@
 
 import React from 'react';
 
-import { InternalLog } from '../../InternalLog';
-import { SdkVerbosity } from '../../SdkVerbosity';
-import { getErrorMessage } from '../../errorUtils';
-import { DdSdk } from '../../foundation';
+import { InternalLog } from '../../../InternalLog';
+import { SdkVerbosity } from '../../../SdkVerbosity';
+import { getErrorMessage } from '../../../errorUtils';
+import { DdSdk } from '../../../foundation';
 
 import { DdEventsInterceptor } from './DdEventsInterceptor';
 import type EventsInterceptor from './EventsInterceptor';
