+++ packages/core/src/rum/instrumentation/interactionTracking/EventsInterceptor.tsx
