+++ packages/core/src/rum/instrumentation/interactionTracking/NoOpEventsInterceptor.tsx
