+++ packages/core/src/rum/instrumentation/interactionTracking/ShallowObjectEqualityChecker.ts
